LPS-145031 SF, remove unnecessary unboxing

+++ portal-kernel/src/com/liferay/portal/kernel/dao/db/DBInspector.java
@@ -298,8 +298,8 @@ public class DBInspector {
 
 		Integer dataTypeSize = _getSQLVarcharSize(columnType);
 
-		if ((dataTypeSize != null) && (dataTypeSize.intValue() != -1)) {
-			return dataTypeSize.intValue();
+		if ((dataTypeSize != null) && (dataTypeSize != -1)) {
+			return dataTypeSize;
 		}
 
 		return -1;
